Update README.md
This Bash script contains two important scripts that will help you to automate the Linux installation process. The first script validates the RAM of the local server and central server. The minimum required RAM for the local server is 4GB, and the central server requires a minimum of 6GB. If the RAM is less than the required amount, it will fail the validation. The second script automates the installation process and saves a lot of time and effort.

Using this script, you can easily check whether your Linux system meets the installation requirements or not. This can help you to avoid any errors or issues during the installation process. Additionally, the script can automate the installation process, which can save you a lot of time and effort.

Overall, this Bash script is a handy tool for anyone who wants to install Linux on their system quickly and easily.